+++ firmware/src/DynamicMultiplexer.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <ArduinoJson.h>
+
+class DynamicMultiplexer {
+ public:
+  static void configureProcessorTopology(JsonObjectConst unit);
+  static void resetAll();
+};